+++ packages/component-library-react/src/components.ts
@@ -23,5 +23,7 @@ export const PlmgSidebarItem = /*@__PURE__*/createReactComponent<JSX.PlmgSidebar
 export const PlmgSlider = /*@__PURE__*/createReactComponent<JSX.PlmgSlider, HTMLPlmgSliderElement>('plmg-slider');
 export const PlmgStatus = /*@__PURE__*/createReactComponent<JSX.PlmgStatus, HTMLPlmgStatusElement>('plmg-status');
 export const PlmgSvgIcon = /*@__PURE__*/createReactComponent<JSX.PlmgSvgIcon, HTMLPlmgSvgIconElement>('plmg-svg-icon');
+export const PlmgTab = /*@__PURE__*/createReactComponent<JSX.PlmgTab, HTMLPlmgTabElement>('plmg-tab');
+export const PlmgTabs = /*@__PURE__*/createReactComponent<JSX.PlmgTabs, HTMLPlmgTabsElement>('plmg-tabs');
 export const PlmgTextInput = /*@__PURE__*/createReactComponent<JSX.PlmgTextInput, HTMLPlmgTextInputElement>('plmg-text-input');
 export const PlmgTooltip = /*@__PURE__*/createReactComponent<JSX.PlmgTooltip, HTMLPlmgTooltipElement>('plmg-tooltip');